1.2.5.3 Se elimino el app example y se creo la conexion con home
