+++ home/templates/home/index.html
@@ -6,6 +6,6 @@
     <title>Inicio</title>
 </head>
 <body>
-    Inicio
+    
 </body>
 </html>
